requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.7.0
APScheduler==3.6.3
attrs==24.3.0
blinker==1.9.0
cachetools==4.2.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
python-telegram-bot==21.9
openai==0.27.8
pytz==2024.2
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
tzlocal==5.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3